requirements.txt
backcall==0.2.0
decorator==5.1.0
ipython==7.28.0
jedi==0.18.0
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.21
psycopg2-binary==2.9.1
ptyprocess==0.7.0
Pygments==2.10.0
traitlets==5.1.0
wcwidth==0.2.5